server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pandas==2.1.4
numpy==1.24.3
google-generativeai==0.3.2
httpx==0.25.2
pydantic-settings==2.1.0
